+++ extend/bajo/hook/dobo.sumba-user@after-update-record.js
@@ -1,3 +1,14 @@
+export async function clearCache (id, result) {
+  if (!this.app.bajoCache) return
+  const { hash } = this.app.bajoExtra
+  const { clear } = this.app.bajoCache
+  const { get, isEmpty } = this.app.lib._
+  let token = get(result, 'data.token', get(result, 'oldData.token', ''))
+  if (!isEmpty(token)) token = await hash(token)
+  await clear({ key: `dobo|SumbaUser|getUserById|${id}` })
+  await clear({ key: `dobo|SumbaUser|getUserByToken|${token}` })
+}
+
 async function afterUpdateRecord (id, body, rec, options = {}) {
   const { data, oldData } = rec
   const { req } = options
@@ -5,6 +16,9 @@ async function afterUpdateRecord (id, body, rec, options = {}) {
   const t = get(req, 't', this.t)
   const to = `${data.firstName} ${data.lastName} <${data.email}>`
   const source = this.ns
+
+  await clearCache.call(this, id, rec)
+
   let subject
   const payload = { to, subject, data }