+++ extend/waibuMpa/route/your-stuff/profile.js
@@ -3,7 +3,7 @@ const profile = {
   handler: async function (req, reply) {
     const { hash } = this.app.bajoExtra
     const { getRecord } = this.app.waibuDb
-    const options = { forceNoHidden: ['token'], noHook: true, noCache: true, attachment: true, mimeType: true }
+    const options = { forceNoHidden: ['token'], noHook: true, noCache: true, attachment: true, mimeType: true, formatValue: true, retainOriginalValue: true }
     const resp = await getRecord({ model: 'SumbaUser', req, id: req.user.id, options })
     const form = resp.data
     form.token = await hash(form.salt)